backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
